Move test_device_parallelism to gpu_test.py
